+++ data/model/retirement_focus_type.py
@@ -1,0 +1,6 @@
+import enum
+
+
+class RetirementFocusType(enum.Enum):
+    LEGACY = "LEGACY"
+    PROLONGED_LIFE = "PROLONGED_LIFE"